ai-agent-service/requirements.txt
# Python dependencies for AI Agent Service
fastapi==0.115.0
uvicorn[standard]==0.32.0

# LangChain ecosystem (compatible versions)
langchain==0.3.0
langchain-core==0.3.84
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-text-splitters==0.3.11

httpx==0.28.0
pydantic==2.10.0
pydantic-settings==2.6.0
pyyaml==6.0.2
python-multipart==0.0.20

# Multi-Agent Framework (NEW)
langgraph==0.2.28  # Updated version compatible with langchain-core 0.3.x

# RAG Knowledge Base (NEW)
chromadb==0.5.0
langchain-chroma==0.1.4  # LangChain ChromaDB integration

# Video Analysis Tools (NEW)
ffmpeg-python==0.2.0  # FFmpeg Python wrapper

# OCR Tools (NEW)
pytesseract==0.3.10
pillow==11.0.0  # PIL图像处理

# PPT Analysis Tools (NEW)
python-pptx==0.6.23  # PowerPoint文件处理

# Speech Recognition (NEW)
openai-whisper==20231117  # OpenAI Whisper语音识别

# Async support
aiofiles==24.1.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2  # YAML配置文件解析（skills.yaml）